# backend/swagger.yaml
openapi: "3.0.0"
info:
  title: McMaster CanIDS
  version: 2.0.0
paths:
...
  /api/group/delete:
    post:
      summary: Delete existing group
      description: |
        Delete existing
        
        Restrictions:
          - `superuser` is the only class of accounts allowed to delete groups.

      tags:
      - Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  description: |
                    UUID of group to delete
              required:
                - uuid
            example: {
              "uuid": "175dd231-f547-4342-a33e-c45d5e99aa7e"
            }
      responses:
        '200':
          description: |
            Group successfully updated.
          content:
            application/json:
              example: {
                "success": true,
                "message": "Successfully updated group."
              }
        '400':
          description: |
            Request parameters are not valid.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Bad request format."
              }
        '401':
          description: |
            User is not authenticated
        '403':
          description: |
            User does not have permissions to perform requested actions.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Only a superuser can create a new group."
              }
        '500':
          description: |
            Internal server error
...
  /api/dashboard/get:
    get:
      summary: Fetch dashboard
      description: |
        Returns the group dashboard. If a dashboard does not exist, automatically pre-loads views and the dashboard.
        
        `superuser` may specify `group` query parameter to fetch dashboard of other group.
      tags:
      - Dashboard
      parameters: 
        - in: query
          name: group
          schema:
            type: string
          description: |
            UUID of group to get dashboard for.
          required: false
      responses:
        '200':
          description: |
            Dashboard configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  uuid:
                    type: string
                    description: |
                      Dashboard UUID
                  group:
                    type: string
                    description: |
                      UUID of group dashboard belongs to
                  name:
                    type: string
                    description: |
                      Dashboard name
                  views:
                    type: array
                    description: |
                      List of views in dashboard
                    items:
                      type: string
                      description: |
                        View UUID in dashboard
                  sizes:
                    type: array
                    description: |
                      List of sizes corresponding to views: `half` or `full` width
                    items:
                      type: string
                      description: |
                        Size corresponding to UUID in dashboard
                required:
                - uuid
                - group
                - name
                - views
                - sizes
              example: {
                "uuid": "7290169e-9bdf-437b-8dd6-456b17341f05",
                "group": "6a7d3a03-5c06-4620-8325-583c8136b17f",
                "name": "McMaster University Dashboard",
                "views": [
                  "70884425-0cab-4840-a99e-ee397a9dbc71",
                  "b93b03f2-64ce-45cd-8737-9f99ed885253"
                ],
                "sizes": [
                  "half",
                  "half"
                ]
              }
        '401':
          description: |
            User is not authenticated
        '500':
          description: |
            Internal server error
  /api/dashboard/update:
    post:
      summary: Update dashboard
      description: |
        Update an existing view.
        
        Restrictions:
          - `admin` can update their group's dashboard
          - `superuser` can update any group's dashboard
      tags:
      - Dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  description: |
                    Dashboard UUID
                group:
                  type: string
                  description: |
                    UUID of group dashboard belongs to
                name:
                  type: string
                  description: |
                    Dashboard name
                views:
                  type: array
                  description: |
                    List of views in dashboard
                  items:
                    type: string
                    description: |
                      View UUID in dashboard
                sizes:
                  type: array
                  description: |
                    List of sizes corresponding to views: `half` or `full` width
                  items:
                    type: string
                    description: |
                      Size corresponding to UUID in dashboard
              required:
              - uuid
              - group
              - name
              - views
              - sizes
            example: {
              "uuid": "7290169e-9bdf-437b-8dd6-456b17341f05",
              "group": "6a7d3a03-5c06-4620-8325-583c8136b17f",
              "name": "McMaster University Dashboard",
              "views": [
                "70884425-0cab-4840-a99e-ee397a9dbc71",
                "b93b03f2-64ce-45cd-8737-9f99ed885253"
              ],
              "sizes": [
                "full",
                "full"
              ]
            }
      responses:
        '200':
          description: |
            Dashboard successfully updated.
          content:
            application/json:
              example: {
                "success": true,
                "message": "Dashboard successfully updated."
              }
        '400':
          description: |
            Request parameters are not valid.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Bad request format."
              }
        '401':
          description: |
            User is not authenticated
        '403':
          description: |
            User does not have permissions to perform requested actions.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Admin users can not update foreign dashboard."
              }
        '500':
          description: |
            Internal server error